feat: osco consolidation and simplification
